astro_airflow_mlflow/project_requirements.txt
astro-run-dag 
a2wsgi==1.10.8
aiobotocore==2.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiohttp-retry==2.9.1
aioitertools==0.12.0
aiologic==0.14.0
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.16.1
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
apache-airflow==3.0.1
apache-airflow-core==3.0.1
apache-airflow-providers-common-compat==1.7.0
apache-airflow-providers-common-io==1.6.0
apache-airflow-providers-common-sql==1.27.1
apache-airflow-providers-http==5.3.0
apache-airflow-providers-postgres==6.2.0
apache-airflow-providers-smtp==2.1.0
apache-airflow-providers-standard==1.2.0
apache-airflow-task-sdk==1.0.1
appdirs==1.4.4
argcomplete==3.6.2
asgiref==3.8.1
asyncpg==0.30.0
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
billiard==4.2.1
blinker==1.9.0
boto3==1.37.3
botocore==1.37.3
cachetools==5.5.2
cadwyn==5.3.3
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
configobj==5.0.9
contourpy==1.3.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==45.0.3
cycler==0.12.1
databricks-sdk==0.55.0
Deprecated==1.2.18
dictdiffer==0.9.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
Django==5.2.1
dnspython==2.7.0
docker==7.1.0
dpath==2.2.0
dulwich==0.22.8
dvc==3.59.2
dvc-data==3.16.10
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-s3==3.2.0
dvc-studio-client==0.21.0
dvc-task==0.40.2
email_validator==2.2.0
ensure==1.0.4
entrypoints==0.4
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.58.1
frozenlist==1.6.0
fsspec==2025.5.1
funcy==2.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.2
googleapis-common-protos==1.70.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.2
grpcio==1.71.0
gto==1.7.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hydra-core==1.3.2
hyperopt==0.2.7
idna==3.10
importlib_metadata==8.6.1
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kombu==5.5.4
lazy-object-proxy==1.11.0
libcst==1.8.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
methodtools==0.4.7
mlflow==2.22.0
mlflow-skinny==2.22.0
more-itertools==10.7.0
msgspec==0.19.0
multidict==6.4.4
networkx==3.5
numpy==2.2.6
omegaconf==2.3.0
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pendulum==3.1.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
py4j==0.10.9.9
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydot==4.0.0
pygit2==1.18.0
Pygments==2.19.1
pygtrie==2.5.0
PyJWT==2.10.1
pyparsing==3.2.3
python-box==7.3.2
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
retryhttp==1.3.3
rich==14.0.0
rich-argparse==1.7.1
rich-toolkit==0.14.7
rpds-py==0.25.1
rsa==4.9.1
ruamel.yaml==0.18.12
ruamel.yaml.clib==0.2.12
s3fs==2025.5.1
s3transfer==0.11.3
scikit-learn==1.6.1
scipy==1.15.3
scmrepo==3.3.11
seaborn==0.13.2
semver==3.0.4
setproctitle==1.3.6
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
sqltrie==0.11.2
starlette==0.46.2
structlog==25.3.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
threadpoolctl==3.6.0
tomlkit==0.13.2
tqdm==4.67.1
typer==0.16.0
types-PyYAML==6.0.12.20250516
types-requests==2.32.0.20250602
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.4.0
uuid6==2024.7.10
uvicorn==0.34.3
uvloop==0.21.0
vine==5.1.0
voluptuous==0.15.2
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wirerope==1.0.0
wrapt==1.17.2
yarl==1.20.0
zc.lockfile==3.0.post1
zipp==3.22.0
